feat: 클래스 리뷰 관리 목록 조회 추가

+++ src/main/java/com/clanone/onedayclan/clazz/adapter/in/web/response/AdminClassInfoResponse.java
@@ -10,6 +10,6 @@ public class AdminClassInfoResponse {
     private String thumbnailUrl;
     private String className;
     private String classCategory;
-    private int attendanceCount;
+    private long attendanceCount;
     private LocalDateTime startAt;
 }
