+++ src/main/java/com/clanone/onedayclan/clazz/adapter/in/web/response/AdminClassReviewInfoResponse.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class AdminClassReviewInfoResponse {
     private long seq;
-    private int reviewCount;
-    private int reviewScore;
+    private long reviewCount;
+    private double reviewScore;
 }
